each class now has its own file + bug fixing
