+++ thefacetakt/src/main/java/ru/fizteh/fivt/students/thefacetakt/TwitterStream/InvalidLocationException.java
@@ -1,0 +1,13 @@
+package ru.fizteh.fivt.students.thefacetakt.TwitterStream;
+
+/**
+ * Created by thefacetakt on 23.09.15.
+ */
+class InvalidLocationException extends Exception {
+    InvalidLocationException() {
+    }
+
+    InvalidLocationException(String message) {
+        super(message);
+    }
+}
