+++ thefacetakt/src/main/java/ru/fizteh/fivt/students/thefacetakt/TwitterStream/Location.java
@@ -1,0 +1,32 @@
+package ru.fizteh.fivt.students.thefacetakt.TwitterStream;
+
+/**
+ * Created by thefacetakt on 23.09.15.
+ */
+class Location {
+    private double latitude;
+    private double longitude;
+    private String name;
+
+    public double getLatitude() {
+        return latitude;
+    }
+
+    public double getLongitude() {
+        return longitude;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String newName) {
+        this.name = newName;
+    }
+
+
+    Location(double newLatitude, double newLongitude) {
+        this.latitude = newLatitude;
+        this.longitude = newLongitude;
+    }
+}
