+++ thefacetakt/src/main/java/ru/fizteh/fivt/students/thefacetakt/TwitterStream/TwitterStream.java
@@ -7,400 +7,8 @@ package ru.fizteh.fivt.students.thefacetakt.TwitterStream;
 import com.beust.jcommander.*;
 import twitter4j.*;
 
-import java.io.*;
-import java.net.MalformedURLException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URL;
 import java.text.SimpleDateFormat;
 import java.util.*;
-
-class Location {
-    private double latitude;
-    private double longitude;
-    private String name;
-
-    public double getLatitude() {
-        return latitude;
-    }
-
-    public double getLongitude() {
-        return longitude;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String newName) {
-        this.name = newName;
-    }
-
-
-    Location(double newLatitude, double newLongitude) {
-        this.latitude = newLatitude;
-        this.longitude = newLongitude;
-    }
-}
-
-class InvalidLocationException extends Exception {
-    InvalidLocationException() {
-    }
-
-    InvalidLocationException(String message) {
-        super(message);
-    }
-}
-
-class QueryLimitException extends Exception {
-    QueryLimitException() {
-    }
-
-    QueryLimitException(String message) {
-        super(message);
-    }
-}
-
-class PlaceLocationResolver {
-    static final String LOCATION_DEFINITION_ERROR
-            = "Problem while location definition";
-
-    private Map<String, Location> cache
-            = new HashMap<String, Location>();
-    private String googleMapsKey;
-    private String yandexMapsKey;
-
-    PlaceLocationResolver() {
-        try (BufferedReader mapsKeyFile = new BufferedReader(
-                new FileReader("src/main/resources/geo.properties"))) {
-            googleMapsKey = mapsKeyFile.readLine();
-            yandexMapsKey = mapsKeyFile.readLine();
-        } catch (IOException e) {
-            System.err.println("Something went terribly wrong: no maps "
-                    + "key found");
-            System.exit(1);
-        }
-    }
-
-
-    private Location resolvePlaceLocationGoogle(String nameOfLocation)
-            throws InvalidLocationException, QueryLimitException {
-        int numberOfTries = 0;
-
-        do {
-            URL googleMapsURL = null;
-            try {
-                try {
-                    URI uri = new URI("https",
-                            "maps.googleapis.com",
-                            "/maps/api/geocode/json",
-                            "address=" + nameOfLocation + "&key="
-                                    + googleMapsKey,
-                            null);
-                    googleMapsURL = uri.toURL();
-                } catch (URISyntaxException e) {
-                    e.printStackTrace();
-                }
-            } catch (MalformedURLException e) {
-                System.err.println(e.getMessage());
-                ++numberOfTries;
-                continue;
-            }
-
-            try {
-                BufferedReader in = new BufferedReader(new InputStreamReader(
-                        googleMapsURL.openStream()));
-
-                String currentInfo;
-                StringBuilder responseStrBuilder = new StringBuilder();
-                while ((currentInfo = in.readLine()) != null) {
-                    responseStrBuilder.append(currentInfo);
-                }
-
-                JSONObject locationInfo =
-                        new JSONObject(responseStrBuilder.toString());
-
-                String status = locationInfo.getString("status");
-                if (status.equals("OVER_QUERY_LIMIT")) {
-                    throw new QueryLimitException("Google");
-                }
-                if (!status.equals("OK")) {
-                    throw new InvalidLocationException("Unknown place");
-                }
-
-                locationInfo = locationInfo
-                        .getJSONArray("results")
-                        .getJSONObject(0)
-                        .getJSONObject("geometry")
-                        .getJSONObject("location");
-
-                return new Location(
-                        Double.parseDouble(locationInfo.getString("lat")),
-                        Double.parseDouble(locationInfo.getString("lng")));
-            } catch (IOException | JSONException e) {
-                System.out.println(e.getMessage());
-                ++numberOfTries;
-                continue;
-            }
-        }
-        while (numberOfTries < TwitterStream.MAX_NUMBER_OF_TRIES);
-        System.err.println(LOCATION_DEFINITION_ERROR);
-        System.exit(1);
-        return null;
-    }
-
-    private Location resolvePlaceLocationYandex(String nameOfLocation)
-            throws InvalidLocationException {
-        int numberOfTries = 0;
-
-        do {
-            URL yandexMapsURL = null;
-            try {
-                try {
-                    URI uri = new URI("https",
-                            "geocode-maps.yandex.ru",
-                            "/1.x/",
-                            "geocode=" + nameOfLocation
-                                    + "&key=" + yandexMapsKey
-                                    + "&format=json",
-                            null);
-                    yandexMapsURL = uri.toURL();
-                } catch (URISyntaxException e) {
-                    e.printStackTrace();
-                }
-            } catch (MalformedURLException e) {
-                System.err.println(e.getMessage());
-                ++numberOfTries;
-                continue;
-            }
-
-            try {
-                assert yandexMapsURL != null;
-
-                BufferedReader in = new BufferedReader(new InputStreamReader(
-                        yandexMapsURL.openStream()));
-
-                String currentInfo;
-                StringBuilder responseStrBuilder = new StringBuilder();
-                while ((currentInfo = in.readLine()) != null) {
-                    responseStrBuilder.append(currentInfo);
-                }
-
-                JSONObject locationInfo =
-                        new JSONObject(responseStrBuilder.toString())
-                                .getJSONObject("response")
-                                .getJSONObject("GeoObjectCollection");
-
-                if (Integer.parseInt(locationInfo
-                        .getJSONObject("metaDataProperty")
-                        .getJSONObject("GeocoderResponseMetaData")
-                        .getString("found")) == 0) {
-                    throw new InvalidLocationException("Unknown place");
-                }
-
-                locationInfo = locationInfo
-                        .getJSONArray("featureMember")
-                        .getJSONObject(0)
-                        .getJSONObject("GeoObject")
-                        .getJSONObject("Point");
-                String[] coordinates =
-                        locationInfo.getString("pos").split(" ");
-
-                return new Location(
-                        Double.parseDouble(coordinates[1]),
-                        Double.parseDouble(coordinates[0]));
-            } catch (IOException | JSONException e) {
-                System.out.println(e.getMessage());
-                ++numberOfTries;
-            }
-        }
-        while (numberOfTries < TwitterStream.MAX_NUMBER_OF_TRIES);
-        System.err.println(LOCATION_DEFINITION_ERROR);
-        System.exit(1);
-        return null;
-    }
-
-    public Location resolvePlaceLocation(String nameOfLocation)
-            throws InvalidLocationException {
-        nameOfLocation = nameOfLocation.trim();
-
-        if (nameOfLocation.length() == 0) {
-            throw new InvalidLocationException("empty address");
-        }
-
-        if (!cache.containsKey(nameOfLocation)) {
-
-            try {
-                Location result
-                        = resolvePlaceLocationGoogle(nameOfLocation);
-                cache.put(nameOfLocation, result);
-            } catch (QueryLimitException e) {
-                try {
-                    Location result
-                            = resolvePlaceLocationYandex(nameOfLocation);
-                    cache.put(nameOfLocation, result);
-                } catch (InvalidLocationException ie) {
-                    cache.put(nameOfLocation, null);
-                }
-            } catch (InvalidLocationException e) {
-                cache.put(nameOfLocation, null);
-            }
-
-        }
-
-        if (cache.get(nameOfLocation) == null) {
-            throw new InvalidLocationException("Unknown place");
-        }
-        return cache.get(nameOfLocation);
-    }
-
-    public Location resolveCurrentLocation() {
-        int numberOfTries = 0;
-
-        do {
-            URL whatIsMyCity;
-            try {
-                whatIsMyCity = new URL("http://ipinfo.io/json");
-            } catch (MalformedURLException e) {
-                ++numberOfTries;
-                continue;
-            }
-
-            try {
-                BufferedReader in = new BufferedReader(new InputStreamReader(
-                        whatIsMyCity.openStream()));
-
-                String currentInfo;
-                StringBuilder responseStrBuilder = new StringBuilder();
-                while ((currentInfo = in.readLine()) != null) {
-                    responseStrBuilder.append(currentInfo);
-                }
-
-                JSONObject locationInfo =
-                        new JSONObject(responseStrBuilder.toString());
-
-                String[] coordinates = locationInfo
-                        .getString("loc").split(",");
-
-                Location result = new Location(
-                        Double.parseDouble(coordinates[0]),
-                        Double.parseDouble(coordinates[1]));
-                result.setName(locationInfo.getString("city"));
-                return result;
-            } catch (IOException | JSONException e) {
-                ++numberOfTries;
-            }
-        }
-        while (numberOfTries < TwitterStream.MAX_NUMBER_OF_TRIES);
-
-        System.err.println(LOCATION_DEFINITION_ERROR);
-        System.exit(1);
-        return null;
-    }
-}
-
-class JCommanderSetting {
-
-    static final String DEFAULT_LOCATION = "nearby";
-
-    @Parameter(names = {"--query", "-q"},
-            description = "query or keywords for stream",
-            variableArity = true)
-    private List<String> queries = new ArrayList<>();
-
-    @Parameter(names = {"--place", "-p"}, description = "location")
-    private String location = DEFAULT_LOCATION;
-
-    @Parameter(names = {"--stream", "-s"},
-            description = "is stream enabled")
-    private boolean stream = false;
-
-    @Parameter(names = {"--hideRetweets"}, description = "hide retweets")
-    private boolean hideRetweets = false;
-
-    @Parameter(names = {"--limit", "-l"}, description = "limit of tweets")
-    private Integer limit = Integer.MAX_VALUE;
-
-    @Parameter(names = {"--help", "-h"},
-            description = "show help", help = true)
-    private boolean help;
-
-    public List<String> getQueries() {
-        return queries;
-    }
-    public String getLocation() {
-        return location;
-    }
-
-    public boolean isStream() {
-        return stream;
-    }
-
-    public boolean isHideRetweets() {
-        return hideRetweets;
-    }
-
-    public Integer getLimit() {
-        return limit;
-    }
-
-    public boolean isHelp() {
-        return help;
-    }
-}
-
-class Declenser {
-    static final int FIRST_IMPORTANT_MOD = 100;
-    static final int SECOND_IMPORTANT_MOD = 10;
-
-    static final int FIRST_CASE_LEFT_BORDER = 11;
-    static final int FIRST_CASE_RIGHT_BORDER = 19;
-
-    static final int SECOND_CASE = 1;
-
-    static final int THIRD_CASE_LEFT_BORDER = 2;
-    static final int THIRD_CASE_RIGHT_BORDER = 4;
-
-    static String declensionScheme(long n, String[] cases) {
-        n = n % FIRST_IMPORTANT_MOD;
-        if (FIRST_CASE_LEFT_BORDER <= n
-                && n <= FIRST_CASE_RIGHT_BORDER) {
-            return cases[0];
-        }
-
-        n = n % SECOND_IMPORTANT_MOD;
-        if (n == SECOND_CASE) {
-            return cases[1];
-        }
-        if (THIRD_CASE_LEFT_BORDER <= n
-                && n <= THIRD_CASE_RIGHT_BORDER) {
-            return cases[2];
-        }
-        return cases[0];
-    }
-
-    static final String[] RETWEETS_CASES = {"ретвитов", "ретвит", "ретвита"};
-    static String retweetDeclension(long n) {
-
-        return declensionScheme(n, RETWEETS_CASES);
-    }
-
-    static final String[] MINUTES_CASES = {"минут", "минуту", "минуты"};
-    static String minutesDeclension(long n) {
-
-        return declensionScheme(n, MINUTES_CASES);
-    }
-
-    static final String[] HOURS_CASES = {"часов", "час", "часа"};
-    static String hoursDeclension(long n) {
-        return declensionScheme(n, HOURS_CASES);
-    }
-
-    static final String[] DAYS_CASES = {"дней", "день", "дня"};
-    static String daysDeclension(long n) {
-        return declensionScheme(n, DAYS_CASES);
-    }
-}
 
 
 public class TwitterStream {
@@ -422,7 +30,7 @@ public class TwitterStream {
 
     static Location resolveLocation(String passedLocation) {
         Location result = null;
-        if (passedLocation == JCommanderSetting.DEFAULT_LOCATION) {
+        if (passedLocation.equals(JCommanderSetting.DEFAULT_LOCATION)) {
             result = geoResolver.resolveCurrentLocation();
         } else {
             try {
